Updated openstack/openstack Project: openstack/nova  3b43f769c530891891c648d6150395ef4bafe4c5
VMware: Make DatastorePath hashable

Make DatastorePath work properly when used as a key value in a dict or
set.

This change specifically enables a subsequent change in the fake
driver, which stores DatastorePath objects in a set. However, it's
probably a good idea in general as I spent quite some time trying to
work out why it didn't work when I hit it.

Change-Id: Ia5519b84bc47ea1e0a5a70f9368806f762d12db5