Updated openstack/openstack Project: openstack/oslo.concurrency  5d1461204d1cce680378d5b7c6c2b17f19b7f4b0
Make lockutils main() a console entry point

This is more lib-ish than calling python -m on the module.

As part of this change, I also improved the unit tests for this code.
Before we weren't unsetting OSLO_LOCK_PATH before calling the main
function, so we had no way of knowing if it was being set correctly.

I also added a test case to verify return value propagation and
removed a private method that was never called.

Change-Id: I6c35b5409bf567767c5c71b9041dd7f7a012255d