Updated openstack/openstack Project: openstack/oslo.concurrency  394fbf4f046bcaa4e0591238b9fc06f1769e6ac5
Remove hard dep on eventlet

We don't want to force users of this library to pull in eventlet,
so conditionally import it in processutils and only use subprocess
from eventlet if the stdlib is monkey patched.  eventlet doesn't
monkey patch subprocess so we can't rely on it to do that for us.

Also uses sleep from the stdlib time module since it will be
monkey patched when eventlet is in use and the right version will
be used.

Change-Id: I0df528d5ad0d67d6bbc0e7c217ef2d7fa45cce9a