Updated openstack/openstack Project: openstack/glance_store  c87338351b02fe0fd430c3e43b185e3b11c1cb47
Make rbd store's pool handling more universal

Currently we ignore the pool part of the location throughout the rbd store code.
If there is a pool specified, use that. Otherwise we can still fall back to the
configured pool.

This is a required change if we want to support ephemeral disk snapshotting later on
as in that scenario the ephemeral pool might be a different pool from images pool,
yet we're going to need to reference the snapshot of a disk in the ephemeral pool.

Change-Id: Ie415667a809975948c8cfb71ec63a0905995fa67
Closes-Bug: 1368128