Updated openstack/openstack Project: openstack/oslo.vmware  1c8e851887519f141e5b3971e06d4e9468fb8b1b
VimExceptions need to support i18n objects

When lazy is enabled the i18n translation object does not support
str() which causes failures like:
  UnicodeError: Message objects do not support str() because they may
  contain non-ascii characters. Please use unicode() or translate()
  instead.

Removed str() in __init__ and updated __str__ and __unicode__
in VimException and VimFaultsException to correctly handle this object.
Also updated a couple of test cases to use an i18n translation object.

Closes-Bug: #1374210
Change-Id: Id55b5deb5535ba5352b449a3b78099aa912a0b54