Updated openstack/openstack Project: openstack/keystonemiddleware  e77a7a225b0902da47fc4acd643d76ebd77e68d1
Support service user and project in non-default domain

The domain for the service user and project couldn't be configured,
so the auth_token middleware always used the v2 API and the default
domain. With this change, the deployer can configure the domain for
the service user and project and the auth_token middleware will then
use the v3 API to get the service token.

DocImpact
Closes-Bug: #1372142

Change-Id: I4f65adb523a591816a6cd807070d92ce6e414a1b