Updated openstack/openstack Project: openstack/ceilometer  f7e2698b173c0f8275177f77661fd949f27979cc
Per-source separation of static resources & discovery

Previously, the amalgamation of static resources and discovery
extensions defined for all matching pipeline sources were passed
to each pollster on each polling cycle.

This led to unintended duplication of the samples emitted when
an individual pollster matched several sources.

Now, we relate the static resources and discovery extensions to
the originating sources and only pass these when a pollster is
traversed in the context of that source.

Similarly, sinks are now related to the originating source and
samples are only published over the sinks corresponding to the
current sources.

Closes-Bug: #1357869

Change-Id: Ie973625325ba3e25c76c90e4792eeaf466ada657