Updated openstack/openstack Project: openstack/heat  fdb2bf8059361b2bb335eee11c6873a8a103e209
Pass the correct matching_metadata to Ceilometer

When using the Cloud Watch Resource Template don't really have
the json functions to prepend a prefix to the matching_metadata.

The issue here is the nova metadata get converted by Ceilometer from
"metering" to "user_metadata" so depending on the meter name we need to
search for different metadata.

Change-Id: Id3577c8af3f0433577b1ec7b5068585a74f4c3cc