Updated openstack/openstack Project: openstack/nova  d41301c686eca98a23d6868e334007d95a3ba2ed
Log original error when attaching volume fails

We have a race in the gate to attach an encrypted volume and the only
thing we see in the logs is the DeviceIsBusy error which makes this hard
to debug.

This change logs the original error so we can dig deeper into the root
cause.

Related-Bug: #1348204

Change-Id: I7d0f2571e80a4e55133f823d2a04feaf4dddf2e5