Updated openstack/openstack Project: openstack/glance_store  d90a3acdac1f7ff4adbdc3c1b2b5f4439a69a9a8
Recover from errors while deleting image segments

When Glance attempts to delete an image, which is stored as a
segmented object, if an error occurs while deleting one of the
segments, delete operation fails while leaving behind rest of
the segments.
This patch enables Glance to recover from one or more errors during
segment deletion and proceed with deleting other segments. However,
this patch doesn't do anything with the failed segments themselves.
With this patch it'd be possible that even after an image is
deleted, certain chunks are left behind in the store.

Change-Id: Ib3df99b13abb6aa1f579c1dce9e2b64d6e72da91
Closes-Bug: 1256364