Updated openstack/openstack Project: openstack/taskflow  7a2c40292207fdafc61834fb10d0ef50f79b5b6c
Mention issue with more than one thread and reduce workers

Due to how it appears the filesystem transport in kombu is
not thread-safe we will work around this by not having more
than one worker active at the same time in this example.

Oddly it appears the memory transport is unaffected (but
from looking at the code it doesn't look safe either), this
may just be due to how the python memory model works though.

Part of blueprint more-examples

Change-Id: Idaf04fb1a6a622af292511bbcf25329c9a5aab53