Updated openstack/openstack Project: openstack/taskflow  54acb650791bf195ddc0f156135edeec1a65f287
Adjust the WBE log levels

To conform better with the logging level standards
move away from using LOG.exception when the level
is more appropriately a warning/warn.

Also changes how a message that can not be sent is really
a critical error and should be treated as such (since such an
error affects the overall execution model).

Change-Id: I7cebd882b655958d539be36ce3b4deb75ac4a0b7