Updated openstack/openstack Project: openstack/horizon  e843b37539def276f9b6f6e5d417a588dca4c7bc
Fixes downloading a keypair

This patch fixes the invalid error which user gets when downloading
a keypair that wasn't saved from the keypair download link from the
keypair download page.

When user clicks the download link, i just sending that to a different
url and then deleting the keypair and then again creating it with
same name.

Change-Id: I4f9166f112487a500d23f9cc98b98660be2ebe55
Closes-bug: #1182797