Updated openstack/openstack Project: openstack/requirements  4cdcc24fade33823f4da5308b98693acc31ee8a6
Revert "Add a version to pytz."

This reverts commit 09ab86aeb9c1ce0e63c68b374bd0d95ee711ded8.

pytz was originally unqualified, but then it was given a version
because of their version scheme being invalid and detected as
a pre-release by pip. However now pytz has PEP 440 valid versions
and tooling is getting stricter still. This new specifier is
now fully PEP 440 compatible.

Conflicts:
	requirements.txt

Change-Id: I85363c7897f7541a6a45f9bc9eeeae3d7d86de37