Updated openstack/openstack Project: openstack/nova  c5d9fc5b499cf5725f02158bc2154cf53f39a732
Fallback to legacy live migration if config error

Commit ea7da5152cdca7ba674e2137c3899909995e2287 added a path to using
migrateToURI2 for live migration if the version of libvirt used has the
VIR_DOMAIN_XML_MIGRATABLE flag set.

However, a bug in older versions of libvirt causes the live migration to
fail because it's incorrectly validating the old and new domain xml's
for ABI stability.

Not all distros are running with the patched version of libvirt so add a
check in place such that if we fail live migration on the new path with
VIR_ERR_CONFIG_UNSUPPORTED, assume it's due to this issue and attempt
the legacy migrateToURI call.

Closes-Bug: #1362929
Related-Bug: #1279563

Change-Id: Ie82566121c2ed3a6d55919bc111358f4129cb404