Updated openstack/openstack Project: openstack/cinder  fd4a660f32983653abcc3cbc0f6d39f54aeabed0
Fix race condition in ISCSIConnector disconnect_volume

The list of devices returned by driver.get_all_block_devices() will
sometimes contain broken symlinks as the SCSI device has been deleted
but the udev rule for the symlink has not yet completed.

Adding in a check to os.path.exists() will ensure that we will not
consider the broken symlinks as an “in use” device.

Change-Id: Ibb869e10976f894f9e18e9edec6739c2c3bea68c
Closes-Bug: #1375382