Updated openstack/openstack Project: openstack/trove  08b2bc1b2aa622a90dbd2de3b265ee62dc57d7f2
Make the replication snapshot timeout configurable

There is no way to tell how long the snapshot for replication
will take, and we have no good way to poll for the slave state.
Eventually, we will need to have an intelligent poll (perhaps
based on guest heartbeats), but in the meantime we will have
the the snapshot use a configurable timeout which can be set
as needed, and independently of the agent_call timeouts.

Co-Authored-By: Nikhil Manchanda <SlickNik@gmail.com>
Change-Id: I6316d748e91d1ec3eebe25a14bb43fbfe10db669
Closes-bug: 1362310