Updated openstack/openstack Project: openstack/trove  ba3a43f2c5d3233118f0cb152a299c29e6018d36
Partially address concerns in Couchbase restore strategy

This is a preliminary fix for community review. I expect to be making
some changes as well.

This change addresses some concerns about the Couchbase restore
strategy. In particular, the concerns that are addressed include
string injection in some execute calls, and issues around the
permissions on the secret file used to store the password.

This change set aims to fully address the issue around the password
file but is a partial solution to the string injection
problem. Additional changes are required to completely address
that. Some additional bugs have also been raised to address issues
related to other concerns discovered while fixing this bug. More to
come.

Change-Id: Icd8033caea4866f57f8cb4c5861d21797136ff90
Partial-Bug: #1349939