Updated openstack/openstack Project: openstack/keystone  cebb6f2f9de4e9bc806d24b3841c152e099ffebd
Fix parsing of emulated enabled DN

If a non-default emulated enabled DN is specified in configuration,
the DN structure returned from python-ldap is incorrectly converted
to a string.  This leads to an index error when we attempt to extract
the RDN attribute and value.

This patch removes the incorrect string conversion and instead does
the proper conversion on the RDN strings that we extract.

Change-Id: I8f0c4594cfa9a41e1875870c3eb63fae32c8c041
Resolves-bug: #1375772