Updated openstack/openstack Project: openstack/keystone  fceb76c38c58e44eea537230a3a2dd0cc231bd0c
Handle default string values when using user_enabled_invert

When the user_enabled_invert setting is being used, values returned
from LDAP are ultimately converted to a bool type when we reach the
inversion logic.  If the user_enabled_default value is used due to
no value being returned from LDAP, the type is a string.  This causes
the inversion logic to be evaluated incorrectly, as 'not' will return
False for any non-empty string.  This results in disabled accounts
that should be enabled.

Change-Id: Id7b024c12815748305458ca05fc8f8a6324c1908
Closes-bug: #1376053