Updated openstack/openstack Project: openstack/taskflow  75376428186d7794e900ce0ac4df80bfc70078f4
Adjust on_job_posting to not hold the lock while investigating

To avoid issues when investigating jobs in one thread and having
a dispatcher thread get locked on that same investigation lock avoid
this entirely by not holding onto that lock.

This also adds a small cleanup around the notifier thread pool and
ensures that it actually exists in a useable state (and handles the
exceptions that arise when it doesn't) on submission.

Change-Id: I12e71f029726ec54ec0aadbdf71cc3c7a959f047