Updated openstack/openstack Project: openstack-infra/devstack-gate  89ed5e80a2aee04859c499f08a332f0304facaca
Rename subunit_log.txt to testrepository.subunit

Subunit is not a text protocol - it tunnels itself in stdout, but
that is not the same thing. The .txt extension leads to bad headers
from Apache:

wget -S http://logs.openstack.org/.../logs/subunit_log.txt.gz
HTTP/1.1 200 OK
Date: Sun, 31 Aug 2014 23:40:07 GMT
Server: Apache/2.2.22 (Ubuntu)
Vary: Accept-Encoding
Keep-Alive: timeout=5, max=100
Connection: Keep-Alive
Transfer-Encoding: chunked
Content-Type: text/plain

which are clearly bogus - and worse cause wget and curl to treat this
as a plain text file - so they apply line ending transformations to it
which breaks the contents. Some browsers are smarter when one does
'save link' - but all in all its a poor thing to present a binary
object as text.

Don't be confused by the lack of Content-Encoding on the object - thats a
separate issue.

Change-Id: Iacb501e8615a184424f865561cac18c5aa1da1f0