Updated openstack/openstack Project: openstack/glance  d2028eb6dd3cff2e68bda71b434eebfbf72ef0e2
Fix bad header bug in glance-replicator

When trying to use "glance-replicator livecopy" to replicate the
images from one OpenStack glance to another, it encounters "Bad
header" error for "x-image-meta-virtual-size" header. This is
because this header is not in the whitelist of v1 API headers of
form "x-image-meta-xxx".

Another problem after adding this "x-image-meta-virtual-size" to
header whitelist is that the sqlalchemy fail due to virual_size is
"None", which cannot be converted to int. To fix this problem, we
should avoid converting the None value attribute in GET
/v1/images/detail response to the header of POST /v1/images. I
added the None value check, following what the way in
glance/common/utils.py image_meta_to_http_headers() method.

Closes-Bug: #1373333

Change-Id: I13182c4e3212024a9d5663cefd2951c576961f4f