Updated openstack/openstack Project: openstack/oslo.db  abecb9ea1a987ea87baf243e0bae42c6a7ec7ecc
Unwrap DialectFunctionDispatcher from itself.

The DialectFunctionDispatcher.dispatch_for() decorator method
necessarily returns the dispatcher itself and not the
decorated function, so that the object can continue to be
re-used even if the function name is the same as that of the
dispatcher.   In order to support a single function being
wrapped by the dispatcher multiple times with different
criteria, dispatch_for() will now check for the last function
wrapped and use that.

Change-Id: I331670d9b76ae30e7a666648e7e2d4c72641c9ff
Closes-Bug: #1373568