Updated openstack/openstack Project: openstack-infra/devstack-gate  611d751cb2abd7f94644e71ed47acfeb1e6361ad
Always set DATA_DIR

I noticed this when debugging [1]

devstack is intended to set $DATA_DIR to $DEST/data by default, but
currently it does that *before* reading localrc so it always uses its
default $DEST value of /opt/stack.

The code in devstack-gate.sh directly chowns /opt/stack/data/tempest
for the tempest user; this is required for tempest to be able to write
out lock-files.  This currently works due to the above, but if
devstack is fixed to correctly locate DATA_DIR under DEST, the chowned
path is now wrong.

The issue is made more confusing as the extant code explicitly sets
DATA_DIR if running grenade (why is not really explained in the
original change [2]).

The solution is to always explicitly set DATA_DIR to $BASE/data.
Since this is /opt/stack/data, the existing devstack (which
essentially assumes this path) continues to work.  A fixed devstack
will use the explicitly set path and continue to work.

There is no need to explicitly set SWIFT_DATA_DIR as that appears to
be a work-around for an old bug [3]

[1] https://review.openstack.org/#/c/120085/
[2] Id3e7e6d7b44fb715db3bb15600473f776bc87440
[3] https://bugs.launchpad.net/devstack/+bug/1162303

Change-Id: Ic8948bbed51dd90e6c350785a47d539efd589e64