Updated openstack/openstack Project: openstack/cinder  873cd87b46143de16517ec7a8ca7bc58605bfa8c
Windows SMBFS: Handle volume_name in _qemu_img_info

The volume_name is now parsed to the _qemu_img_info wrapper. As
this method is not prone to security issues because this driver
does not support raw images (at least not yet), we don't have to
perform any checks on the backing image file path.

Thus, this method simply ignores this argument that will be parsed
by the base class methods.

Related-Bug: #1350504

Change-Id: I801a6338250ec2dc631c4058543f7d0088b3e4d4