Updated openstack/openstack Project: openstack/tempest  2844bf9729a409dc96a14e8a9757117c6251e446
Don't parse config file if it doesn't exist

This commit adds a check before we parse the config file to see if the
config file actually exists. If it doesn't we shouldn't try to parse
it. There is still an issue with one scenario test that uses
testscenarios but it doesn't result in an error code being raised by a
test-list so that can be tracked down separately.

Change-Id: Iac31f2edcfbccf8c06d73c6ca01e30c8319c0485
Closes-Bug: #1369118